Typescriptify Messages-Component
